requirements.txt
django==3.1.14
djangorestframework==3.11.2
pytest-django==4.4.0
pytest==6.2.5
pytest-cov==2.12.1
